backend/requirements.txt
Flask
Flask-PyMongo
Flask-CORS
Flask-JWT-Extended
python-dotenv
pymongo[srv]
flask-bcrypt
marshmallow
